feat: Mypage 로딩 추가

+++ src/pages/mypage/Mypage.jsx
@@ -3,10 +3,11 @@ import { useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 // Firebase
 import { updateProfile } from "firebase/auth";
-import { doc, updateDoc, deleteDoc } from "firebase/firestore";
+import { doc, updateDoc, deleteDoc, getDoc } from "firebase/firestore";
 import { auth, firestore } from "../../firebase";
 // Slice
 import { logout, setCurrentUser } from "../../redux/user-slice";
+import { startLoading, stopLoading } from "../../redux/global-slice";
 // Toast
 import { toast } from "react-toastify";
 // Sweet Alert
@@ -23,14 +24,17 @@ const Mypage = () => {
 
   // 닉네임 변경 핸들러
   const modifyNicknameHandler = async (modifiedNickname) => {
+    dispatch(startLoading()); // 로딩 화면 시작
     try {
       await updateProfile(auth.currentUser, { displayName: modifiedNickname }); // 1. firebase authentication에 변경된 닉네임 적용
-      const docRef = doc(firestore, "users", docId); // firestore의 users 컬렉션에서 현재 로그인된 유저의 docId에 해당하는 유저 docRef 가져오기
+      const docRef = doc(firestore, "users", docId); // firestore의 users 컬렉션에서 현재 로그인된 유저의 docId에 해당하는 유저 docRef
       await updateDoc(docRef, { nickname: modifiedNickname }); // 2. firestore의 users 컬렉션에 로그인된 유저의 변경된 닉네임 적용
       dispatch(setCurrentUser({ ...auth.currentUser, docId })); // 3. Redux에 변경된 닉네임까지 적용된 현재 로그인된 유저 정보 저장
-      toast.success("닉네임을 성공적으로 변경했습니다"); // 닉네임 변경 성공 토스트
+      toast.success("닉네임 변경 성공"); // 닉네임 변경 성공 토스트
     } catch (error) {
       console.error(error);
+    } finally {
+      dispatch(stopLoading()); // 로딩 화면 종료
     }
   };
 
@@ -39,18 +43,35 @@ const Mypage = () => {
     const title = "정말로 탈퇴하시겠습니까?";
     const result = await MySwal.fire({ ...swalOptions, title });
     if (!result.isConfirmed) return;
-    // toast.warning("팀장으로 있는 동안은 탈퇴할 수 없습니다");
-    // toast.warning("소속된 팀이 존재하는 동안은 탈퇴할 수 없습니다")
+    dispatch(startLoading()); // 로딩 화면 시작
     try {
-      await auth.currentUser.delete(); // 1. firebase authentication에서 현재 로그인된 유저 삭제
-      const docRef = doc(firestore, "users", docId); // firestore의 users 컬렉션에서 현재 로그인된 유저의 docId에 해당하는 유저 docRef 가져오기
-      await deleteDoc(docRef); // 2. firestore의 users 컬렉션에서 현재 로그인된 유저 삭제
-      dispatch(logout()); // 3. Redux에 현재 로그인된 유저 정보 삭제
-      toast.success("회원 탈퇴를 완료했습니다"); // 회원 탈퇴 성공 토스트
+      const docRef = doc(firestore, "users", docId); // firestore의 users 컬렉션에서 현재 로그인된 유저의 docId에 해당하는 유저 docRef
+      const documentSnapshot = await getDoc(docRef); // 1. docRef의 documentSnapshot 가져오기
+      const { teams } = documentSnapshot.data(); // 해당 유저가 소속된 팀 목록 teams
+      if (teams.length > 0) {
+        toast.warning("소속된 팀이 있어 탈퇴 불가"); // 2. 본인이 팀장이거나 팀원으로 소속된 팀이 있으면 탈퇴 불가
+        return;
+      }
+      await deleteDoc(docRef); // 3. firestore의 users 컬렉션에서 현재 로그인된 유저 삭제
+      await auth.currentUser.delete(); // 4. firebase authentication에서 현재 로그인된 유저 삭제
+      dispatch(logout()); // 5. Redux에 현재 로그인된 유저 정보 삭제
+      toast.success("회원 탈퇴 완료"); // 회원 탈퇴 성공 토스트
       navigate("/", { replace: true }); // /로 리다이렉트하고 뒤로가기 방지
     } catch (error) {
       console.error(error);
+    } finally {
+      dispatch(stopLoading()); // 로딩 화면 종료
     }
+  };
+
+  // 비밀번호 변경 핸들러
+  const modifyPasswordHandler = () => {
+    toast.warning("현재 지원하지 않는 기능");
+  };
+
+  // 깃 아이디, 토큰 변경 핸들러
+  const modifyGitHandler = () => {
+    toast.warning("현재 지원하지 않는 기능");
   };
 
   return (
@@ -69,6 +90,8 @@ const Mypage = () => {
       <MypageModifyForm
         modifyNickname={modifyNicknameHandler}
         resign={resignHandler}
+        modifyPassword={modifyPasswordHandler}
+        modifyGit={modifyGitHandler}
       />
     </main>
   );
